backend/requirements.txt
# Machine Intelligence — Backend Dependencies
# Pin versions for reproducible PyInstaller builds

fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3

# Speech-to-text
openai-whisper==20240930

# HTTP
requests==2.32.3
httpx==0.28.1

# Audio processing (Whisper dependency)
soundfile==0.13.1
sounddevice==0.5.1

# Serialport for PlatformIO upload detection
pyserial==3.5

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
